🔨 #34 인터페이스 타입 폴더에 따로 분리, playlistType파일 생성
