+++ src/types/playlistType.d.ts
@@ -1,0 +1,12 @@
+export interface videoListProps {
+  title: string
+  thumbnail: string
+  url: string
+  channelTitle: string
+}
+
+export interface showplaylistProps {
+  playlistId: string
+  title: string
+  thumbnail: string
+}
